+++ govcd/api.go
@@ -601,9 +601,10 @@ func executeRequestCustomErr(pathURL string, params map[string]string, requestTy
 	url, _ := url.ParseRequestURI(pathURL)
 
 	var req *http.Request
-	switch requestType {
-	case http.MethodPost, http.MethodPut:
-
+	switch {
+	// Only send data (and xml.Header) if the payload is actually provided to avoid sending empty body with XML header
+	// (some Web Application Firewalls block requests when empty XML header is set but not body provided)
+	case payload != nil:
 		marshaledXml, err := xml.MarshalIndent(payload, "  ", "    ")
 		if err != nil {
 			return &http.Response{}, fmt.Errorf("error marshalling xml data %s", err)